shell lab: pass all other test except for test13

+++ labs/5. Shell Lab/shlab-handout/tsh.c
@@ -94,7 +94,8 @@ int Sigfillset(sigset_t *set);
 int Sigaddset(sigset_t *set, int signum);
 int Sigprocmask(int how, const sigset_t *set, sigset_t *oldset);
 int Sigsuspend(const sigset_t *mask);
-struct job_t *getjob(struct job_t *jobs, char *id); // get job by pid or jid
+struct job_t *getjob(char **argv, struct job_t *jobs, char *id); 
+// get job by pid or jid
 
 
 
@@ -293,7 +294,18 @@ int builtin_cmd(char **argv)
     } else if (!strcmp(cmd, "jobs")) {
         listjobs(jobs);
         return 1;
+    } else if (!strcmp(cmd, "kill")) {
+        struct job_t *jobp = getjob(argv, jobs, argv[1]);
+        if (jobp == NULL) {
+            printf("%s: No such job\n", argv[1]);
+            return 1;
+        } 
+        kill(jobp->pid, SIGKILL);
+        printf("Job [%d] (%d) terminated by signal %d\n", 
+                jobp->jid, jobp->pid, SIGKILL);
+        return 1;
     }
+
 
     return 0;     /* not a builtin command */
 }
@@ -303,15 +315,23 @@ int builtin_cmd(char **argv)
  */
 void do_bgfg(char **argv) 
 {
-    struct job_t *jobp = getjob(jobs, argv[1]);
+    if (argv[1] == 0) {
+        printf("%s command requires PID or %%jobid argument\n", argv[0]);
+        return ;
+    }
+    struct job_t *jobp = getjob(argv, jobs, argv[1]);
+    if (jobp == NULL) 
+        return ;
     if (!strcmp(argv[0], "bg")) {
         kill(jobp->pid, SIGCONT);
         jobp->state = BG;
+        printf("[%d] (%d) %s", jobp->jid, jobp->pid, jobp->cmdline);
     } else if (!strcmp(argv[0], "fg")) {
         kill(jobp->pid, SIGCONT);
         jobp->state = FG;
         waitfg(jobp->pid);
-    }
+    } 
+        
     return;
 }
 
@@ -320,18 +340,16 @@ void do_bgfg(char **argv)
  */
 void waitfg(pid_t pid)
 {
-    sigset_t sigset;
-    sigset_t oldset;
-
-    Sigemptyset(&sigset);
-    Sigaddset(&sigset, SIGCHLD); 
-    Sigprocmask(SIG_BLOCK, &sigset, &oldset); // Block SIGCHLD
     struct job_t *jobp = getjobpid(jobs, pid);
  
+    sigset_t sigset;
+    sigset_t oldset;
+    sigfillset(&sigset);
+    sigprocmask(SIG_BLOCK, &sigset, &oldset); // block SIGCHLD
     while (jobp->state == FG) {
-        Sigsuspend(&oldset); 
+        Sigsuspend(&oldset);                  // unblock SIGCHLD
     }  
-    Sigprocmask(SIG_SETMASK, &oldset, NULL); // Unblock SIGCHLD
+    sigprocmask(SIG_SETMASK, &oldset, NULL);  // unblock SIGCHLD
     return;
 }
 
@@ -355,12 +373,23 @@ void sigchld_handler(int sig)
     sigset_t oldset;
     Sigfillset(&sigset);
 
-    while ((pid = waitpid(-1, &status, WNOHANG)) > 0) {
-        Sigprocmask(SIG_BLOCK, &sigset, &oldset);
-        deletejob(jobs, pid); 
-        Sigprocmask(SIG_SETMASK, &oldset, NULL);
+    Sigprocmask(SIG_BLOCK, &sigset, &oldset);
+    while ((pid = waitpid(-1, &status, WNOHANG | WUNTRACED)) > 0) {
+        if (WIFEXITED(status)) {
+            deletejob(jobs, pid); 
+        } else if (WIFSIGNALED(status)){
+            printf("Job [%d] (%d) terminate by signal %d\n", 
+                    pid2jid(pid), pid, WTERMSIG(status));
+            deletejob(jobs, pid);
+        } else if (WIFSTOPPED(status)) {
+            struct job_t *jobp = getjobpid(jobs, pid);
+            jobp->state = ST;
+            printf("Job [%d] (%d) stopped by signal %d\n", 
+                    pid2jid(pid), pid, WSTOPSIG(status));
+        }
     }
-    
+
+    Sigprocmask(SIG_SETMASK, &oldset, NULL);
     if (pid < 0 && errno != ECHILD) 
         unix_error("waitpid error");
     return;
@@ -380,9 +409,7 @@ void sigint_handler(int sig)
  
     pid_t fpid = fgpid(jobs);
     if (fpid > 0) {
-        printf("Job [%d] (%d) terminated by signal %d\n", 
-                pid2jid(fpid), fpid, SIGINT);
-        kill(fpid, SIGINT);
+        kill(-fpid, SIGINT);
     }
     Sigprocmask(SIG_SETMASK, &oldset, NULL);
     return;
@@ -402,9 +429,7 @@ void sigtstp_handler(int sig)
 
     pid_t fpid = fgpid(jobs);
     if (fpid > 0) {
-        printf("Job [%d] (%d) terminated by signal %d\n", 
-                pid2jid(fpid), fpid, SIGTSTP);
-        kill(fpid, SIGTSTP);
+        kill(-fpid, SIGTSTP);
     }
     Sigprocmask(SIG_SETMASK, &oldset, NULL);
     return;
@@ -640,7 +665,8 @@ pid_t Fork(void) {
 
 void Execve(char *filename, char **argv, char **envp) {
     execve(filename, argv, envp);
-    unix_error("Execve error");    
+    printf("%s: Command not found\n", filename);
+    exit(0);
 }
 
 int Sigemptyset(sigset_t *set) {
@@ -679,13 +705,29 @@ int Sigsuspend(const sigset_t *mask) {
     return err;
 }
 
-struct job_t *getjob(struct job_t *jobs, char *id) {
+struct job_t *getjob(char **argv, struct job_t *jobs, char *id) {
     if (id[0] == '%') {
         int jid = atoi(id + 1);
-        return getjobjid(jobs, jid);
+        if (jid == 0) {
+            printf("%s: argument must be a PID or %%jobid\n", argv[0]);
+            return NULL;
+        }
+        struct job_t *retjob = getjobjid(jobs, jid);
+        if (retjob == NULL) 
+            printf("%s: No such job\n", id);
+        return retjob;
     }
     else {
         int pid = atoi(id);
-        return getjobpid(jobs, pid);
+        if (pid == 0) {
+            printf("%s: argument must be a PID or %%jobid\n", argv[0]);
+            return NULL;
+        }
+        struct job_t *retjob = getjobpid(jobs, pid);
+        if (retjob == NULL) 
+            printf("(%d): No such process\n", pid);
+
+        return retjob;
+
     }
 }
